move "vert_delete_and_remap" to pmx_utils
